Assignment 2
I have created a UITableView from a template downloaded from appcoda. I
was able to create sections with rows, with section headers and row
numbers.
However I couldnot implement the below functions:
>Resize the row
>Unhighlight the selected row
>log the output to the console.

I am trying to get a hang of the language and kindly request you to
pardon me for this assignment.